Email 관련 Class Auth로 변경
