+++ src/main/java/com/chatbar/domain/auth/presentation/AuthController.java
@@ -4,8 +4,14 @@ import com.chatbar.domain.auth.application.AuthService;
 import com.chatbar.domain.auth.dto.RefreshTokenReq;
 import com.chatbar.domain.auth.dto.SignInReq;
 import com.chatbar.domain.auth.dto.SignUpReq;
+import com.chatbar.domain.email.EmailService;
+import com.chatbar.domain.user.application.UserService;
+import com.chatbar.domain.user.dto.ChangePasswordRes;
+import com.chatbar.domain.user.dto.EmailRes;
+import com.chatbar.domain.user.dto.VerifyRes;
 import com.chatbar.global.config.security.token.CurrentUser;
 import com.chatbar.global.config.security.token.UserPrincipal;
+import com.chatbar.global.payload.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +26,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthService authService;
+
+    private final EmailService emailService;
+
+    private final UserService userService;
 
     //회원가입
     @PostMapping("/sign-up")
@@ -52,5 +62,25 @@ public class AuthController {
              @Valid @RequestBody RefreshTokenReq tokenRefreshRequest){
 
         return authService.refresh(tokenRefreshRequest);
+    }
+
+    @PostMapping("/requestVeri")
+    public ResponseEntity<ApiResponse> requestVerificationCode(@RequestBody EmailRes emailRes) {
+        return emailService.sendVerificationCode(emailRes.getEmail());
+    }
+
+    @PostMapping("/verifyCode")
+    public ResponseEntity<ApiResponse> verifyCode(@RequestBody VerifyRes verifyRes) {
+        return emailService.verifyCode(verifyRes.getEmail(), verifyRes.getCode());
+    }
+
+
+
+    @PostMapping("/changePassword")
+    public ResponseEntity<ApiResponse> changePassword(@RequestBody ChangePasswordRes changePasswordRes) {
+        String email = changePasswordRes.getEmail();
+        String newPassword = changePasswordRes.getNewPassword();
+
+        return userService.updatePasswordByEmail(email, newPassword);
     }
 }
