+++ src/main/java/com/chatbar/domain/user/domain/User.java
@@ -19,6 +19,7 @@ import java.util.EnumSet;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Where(clause = "status = 'ACTIVE'")
+@Table(name = "users")
 public class User extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
